requirements.txt
amqp==5.1.1
asgiref==3.3.1
attrs==22.1.0
billiard==3.6.4.0
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.1
celery==5.2.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.7
django-celery-beat==2.2.1
django-environ==0.9.0
django-templated-mail==1.1.1
django-timezone-field==4.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-spectacular==0.24.2
factory-boy==3.2.1
Faker==11.3.0
firebase-admin==4.5.3
freezegun==1.2.1
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
httplib2==0.19.1
idna==2.10
importlib-metadata==1.7.0
importlib-resources==5.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jsonschema==4.16.0
kombu==5.2.4
MarkupSafe==1.1.1
msgpack==1.0.2
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.2
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.29
proto-plus==1.18.1
protobuf==3.15.6
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.18.1
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
qrcode==6.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
typing_extensions==4.2.0
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
zipp==3.8.0